Moved provenance clause translator into its own file.
